Fixed punlock.
git-svn-id: https://svn.surfnet.nl/svn/ids/sensor/trunk@968 6e3b1d27-7b68-4cde-b5d6-ae3f64e5b54f

+++ sensor/functions/linux.py
@@ -688,8 +688,10 @@ def punlock(fd, file):
         Acts as a proxy for unlocking.
     """
     logging.debugv("functions/linux.py->punlock(fd, file)", [fd, file])
-    fd.close()
-    os.unlink(file)
+    if fd:
+        fd.close()
+    if os.access(file, os.R_OK):
+        os.unlink(file)
 
 def waitInterfaceLink(interface):
     from time import sleep as time_sleep
